Records count int->long fix

+++ api/src/main/java/lsfusion/interop/form/remote/RemoteFormInterface.java
@@ -85,7 +85,7 @@ public interface RemoteFormInterface extends RemoteRequestInterface {
 
     Object getGroupReportData(long requestIndex, long lastReceivedRequestIndex, Integer groupId, FormUserPreferences userPreferences) throws RemoteException;
 
-    int countRecords(long requestIndex, long lastReceivedRequestIndex, int groupObjectID) throws RemoteException;
+    long countRecords(long requestIndex, long lastReceivedRequestIndex, int groupObjectID) throws RemoteException;
 
     Object calculateSum(long requestIndex, long lastReceivedRequestIndex, int propertyID, byte[] columnKeys) throws RemoteException;
